+++ src/index.js
@@ -3,7 +3,6 @@ import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
 import queryString from 'query-string';
 import { hot } from 'react-hot-loader';
-import _ from 'lodash';
 
 import { Route, Switch, IfPermission } from '@folio/stripes/core';
 
@@ -11,115 +10,14 @@ import * as Routes from './routes';
 
 import pkg from '../package';
 import Settings from './settings';
-import PermissionSets from './settings/permissions/PermissionSets';
-import PatronGroupsSettings from './settings/PatronGroupsSettings';
-import AddressTypesSettings from './settings/AddressTypesSettings';
-import ProfilePictureSettings from './settings/ProfilePictureSettings';
-import OwnerSettings from './settings/OwnerSettings';
-import FeeFineSettings from './settings/FeeFineSettings';
-import WaiveSettings from './settings/WaiveSettings';
-import PaymentSettings from './settings/PaymentSettings';
-import CommentRequiredSettings from './settings/CommentRequiredSettings';
-import RefundReasonsSettings from './settings/RefundReasonsSettings';
-import TransferAccountsSettings from './settings/TransferAccountsSettings';
-import CustomFieldsSettingsPane from './settings/CustomFieldsSettings';
+import sections from './settings/sections';
 import {
   NoteCreatePage,
   NoteViewPage,
   NoteEditPage
 } from './views';
 
-const settingsGeneral = [
-  {
-    route: 'perms',
-    label: <FormattedMessage id="ui-users.settings.permissionSet" />,
-    component: PermissionSets,
-    perm: 'ui-users.settings.permsets',
-  },
-  {
-    route: 'groups',
-    label: <FormattedMessage id="ui-users.settings.patronGroups" />,
-    component: PatronGroupsSettings,
-    perm: 'ui-users.settings.usergroups',
-  },
-  {
-    route: 'addresstypes',
-    label: <FormattedMessage id="ui-users.settings.addressTypes" />,
-    component: AddressTypesSettings,
-    perm: 'ui-users.settings.addresstypes',
-  },
-  {
-    route: 'profilepictures',
-    label: <FormattedMessage id="ui-users.settings.profilePictures" />,
-    component: ProfilePictureSettings,
-  },
-  {
-    route: 'custom-fields',
-    label: <FormattedMessage id="ui-users.settings.customFields" />,
-    component: CustomFieldsSettingsPane,
-    perm: 'ui-users.settings.customfields.view',
-  }
-];
-
-const settingsFeefines = [
-  {
-    route: 'owners',
-    label: <FormattedMessage id="ui-users.settings.owners" />,
-    component: OwnerSettings,
-    perm: 'ui-users.settings.feefine',
-  },
-  {
-    route: 'feefinestable',
-    label: <FormattedMessage id="ui-users.settings.manualCharges" />,
-    component: FeeFineSettings,
-    perm: 'ui-users.settings.feefine',
-  },
-  {
-    route: 'waivereasons',
-    label: <FormattedMessage id="ui-users.settings.waiveReasons" />,
-    component: WaiveSettings,
-    perm: 'ui-users.settings.feefine',
-  },
-  {
-    route: 'payments',
-    label: <FormattedMessage id="ui-users.settings.paymentMethods" />,
-    component: PaymentSettings,
-    perm: 'ui-users.settings.feefine',
-  },
-  {
-    route: 'refunds',
-    label: <FormattedMessage id="ui-users.settings.refundReasons" />,
-    component: RefundReasonsSettings,
-    perm: 'ui-users.settings.feefine',
-  },
-  {
-    route: 'comments',
-    label: <FormattedMessage id="ui-users.settings.commentRequired" />,
-    component: CommentRequiredSettings,
-    perm: 'ui-users.settings.feefine',
-  },
-  {
-    route: 'transfers',
-    label: <FormattedMessage id="ui-users.settings.transferAccounts" />,
-    component: TransferAccountsSettings,
-    perm: 'ui-users.settings.transfers',
-  },
-];
-
-export const settingsSections = [
-  {
-    label: <FormattedMessage id="ui-users.settings.general" />,
-    pages: _.sortBy(settingsGeneral, ['label']),
-  },
-  {
-    label: <FormattedMessage id="ui-users.settings.feefine" />,
-    pages: _.sortBy(settingsFeefines, ['label']),
-  },
-];
-
 class UsersRouting extends React.Component {
-  static actionNames = ['stripesHome', 'usersSortByName'];
-
   static propTypes = {
     stripes: PropTypes.shape({
       hasPerm: PropTypes.func.isRequired,
@@ -159,6 +57,8 @@ class UsersRouting extends React.Component {
       }
     }
   }
+
+  static actionNames = ['stripesHome', 'usersSortByName'];
 
   noMatch() {
     const {
@@ -215,7 +115,7 @@ class UsersRouting extends React.Component {
       return (
         <Route path={path} component={Settings}>
           <Switch>
-            {[].concat(...settingsSections.map(section => section.pages))
+            {[].concat(...sections.map(section => section.pages))
               .filter(setting => !setting.perm || stripes.hasPerm(setting.perm))
               .map(setting => <Route path={`${path}/${setting.route}`} key={setting.route} component={setting.component} />)
             }
