+++ src/settings/sections.js
@@ -1,0 +1,105 @@
+import React from 'react';
+import { FormattedMessage } from 'react-intl';
+import { sortBy } from 'lodash';
+
+import PermissionSets from './permissions/PermissionSets';
+import PatronGroupsSettings from './PatronGroupsSettings';
+import AddressTypesSettings from './AddressTypesSettings';
+import ProfilePictureSettings from './ProfilePictureSettings';
+import OwnerSettings from './OwnerSettings';
+import FeeFineSettings from './FeeFineSettings';
+import WaiveSettings from './WaiveSettings';
+import PaymentSettings from './PaymentSettings';
+import CommentRequiredSettings from './CommentRequiredSettings';
+import RefundReasonsSettings from './RefundReasonsSettings';
+import TransferAccountsSettings from './TransferAccountsSettings';
+import CustomFieldsSettingsPane from './CustomFieldsSettings';
+
+
+const settingsGeneral = [
+  {
+    route: 'perms',
+    label: <FormattedMessage id="ui-users.settings.permissionSet" />,
+    component: PermissionSets,
+    perm: 'ui-users.settings.permsets',
+  },
+  {
+    route: 'groups',
+    label: <FormattedMessage id="ui-users.settings.patronGroups" />,
+    component: PatronGroupsSettings,
+    perm: 'ui-users.settings.usergroups',
+  },
+  {
+    route: 'addresstypes',
+    label: <FormattedMessage id="ui-users.settings.addressTypes" />,
+    component: AddressTypesSettings,
+    perm: 'ui-users.settings.addresstypes',
+  },
+  {
+    route: 'profilepictures',
+    label: <FormattedMessage id="ui-users.settings.profilePictures" />,
+    component: ProfilePictureSettings,
+  },
+  {
+    route: 'custom-fields',
+    label: <FormattedMessage id="ui-users.settings.customFields" />,
+    component: CustomFieldsSettingsPane,
+    perm: 'ui-users.settings.customfields.view',
+  }
+];
+
+const settingsFeefines = [
+  {
+    route: 'owners',
+    label: <FormattedMessage id="ui-users.settings.owners" />,
+    component: OwnerSettings,
+    perm: 'ui-users.settings.feefine',
+  },
+  {
+    route: 'feefinestable',
+    label: <FormattedMessage id="ui-users.settings.manualCharges" />,
+    component: FeeFineSettings,
+    perm: 'ui-users.settings.feefine',
+  },
+  {
+    route: 'waivereasons',
+    label: <FormattedMessage id="ui-users.settings.waiveReasons" />,
+    component: WaiveSettings,
+    perm: 'ui-users.settings.feefine',
+  },
+  {
+    route: 'payments',
+    label: <FormattedMessage id="ui-users.settings.paymentMethods" />,
+    component: PaymentSettings,
+    perm: 'ui-users.settings.feefine',
+  },
+  {
+    route: 'refunds',
+    label: <FormattedMessage id="ui-users.settings.refundReasons" />,
+    component: RefundReasonsSettings,
+    perm: 'ui-users.settings.feefine',
+  },
+  {
+    route: 'comments',
+    label: <FormattedMessage id="ui-users.settings.commentRequired" />,
+    component: CommentRequiredSettings,
+    perm: 'ui-users.settings.feefine',
+  },
+  {
+    route: 'transfers',
+    label: <FormattedMessage id="ui-users.settings.transferAccounts" />,
+    component: TransferAccountsSettings,
+    perm: 'ui-users.settings.transfers',
+  },
+];
+
+export default [
+  {
+    label: <FormattedMessage id="ui-users.settings.general" />,
+    pages: sortBy(settingsGeneral, ['label']),
+  },
+  {
+    label: <FormattedMessage id="ui-users.settings.feefine" />,
+    pages: sortBy(settingsFeefines, ['label']),
+  },
+];
